requirements.txt
streamlit
opencv-python
numpy
pytesseract
Pillow